changed to remove subdirectories
